+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file is manually updated. Please file an issue if something is missing.
 
+## v4.6.1, v3.6.1 (2022-11-07)
+
+- **Feature**: Native sharing from the viewer (images only)
+- **Feature**: Deep linking to photos on opening viewer
+- **Feature**: Password protected folder shares ([#165](https://github.com/pulsejet/memories/issues/165))
+- **Feature**: Folders view will now show only folders with photos ([#163](https://github.com/pulsejet/memories/issues/163))
+- Improvements to viewer UX
+- Restore image editor (see v4.6.0)
+
 ## v4.6.0, v3.6.0 (2022-11-06)
 
 - **Brand new photo viewer** with improved touch interface and UX